refactor: Extract ack message from connections to common module
