+++ src/lib/modules/common/messages/CommonMessageType.ts
@@ -1,0 +1,3 @@
+export enum CommonMessageType {
+  Ack = 'did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/notification/1.0/ack',
+}
