+++ src/lib/modules/connections/messages/index.ts
@@ -1,4 +1,3 @@
-export * from './AckMessage';
 export * from './ConnectionInvitationMessage';
 export * from './ConnectionMessageType';
 export * from './ConnectionRequestMessage';
