+++ src/lib/modules/credentials/services/CredentialService.ts
@@ -6,7 +6,7 @@ import { AgentMessage } from '../../../agent/AgentMessage';
 import { LedgerService } from '../../ledger/services/LedgerService';
 import { InboundMessageContext } from '../../../agent/models/InboundMessageContext';
 import { Attachment, AttachmentData } from '../../../decorators/attachment/Attachment';
-import { ConnectionService, ConnectionRecord, AckStatus } from '../../connections';
+import { ConnectionService, ConnectionRecord } from '../../connections';
 import { CredentialRecord } from '../repository/CredentialRecord';
 import { Repository } from '../../../storage/Repository';
 import { JsonEncoder } from '../../../utils/JsonEncoder';
@@ -28,6 +28,7 @@ import {
   ProposeCredentialMessage,
   ProposeCredentialMessageOptions,
 } from '../messages';
+import { AckStatus } from '../../common';
 import { Logger } from '../../../logger';
 import { AgentConfig } from '../../../agent/AgentConfig';
