+++ src/lib/modules/proofs/messages/PresentationAckMessage.ts
@@ -1,6 +1,6 @@
 import { Equals } from 'class-validator';
 
-import { AckMessage, AckMessageOptions } from '../../../modules/connections';
+import { AckMessage, AckMessageOptions } from '../../../modules/common';
 import { PresentProofMessageType } from './PresentProofMessageType';
 
 export type PresentationAckMessageOptions = AckMessageOptions;
